% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nutrients.R
\name{compare_nutrients}
\alias{compare_nutrients}
\title{The function \emph{compare_nutrients} Returns a plot that compares two food items based on their nutritional information.}
\usage{
compare_nutrients(itemA, itemB, measure)
}
\arguments{
\item{itemA}{The first food item, inputed as a string, which you'd like to compare.}

\item{itemB}{The second food item, inputed as a string, which you'd like to compare.}

\item{measure}{The type of nutritional information, inputed as a string, which you'd like to compare the food on.
"calories" = The amount of calories (kcal).
"percent_nutrients" = the percentage of the daily needed intake of nutrients in the food item.
"fat" = amount of fat in grams.
"carbohydrates" = amount of carbohydrates in grams.
"fibre" = amount of fibre in grams.}
}
\value{
A barplot with the two items on the x-axis and the specified nutritional measure on the y-axis.
}
\description{
The function \emph{compare_nutrients} Returns a plot that compares two food items based on their nutritional information.
}
