% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_functions.R
\name{get_nutrients}
\alias{get_nutrients}
\title{\emph{get_nutrients}: get nutritional information for food items.}
\usage{
get_nutrients(food, measure)
}
\arguments{
\item{food}{The food item, inputed as a string, you would like to search for.}

\item{measure}{The type of nutritional information, inputed as a string, per serving of the specified food item.
"calories" = The amount of calories (kcal).
"percent_nutrients" = the percentage of the daily needed intake of nutrients in the food item.
"fat" = amount of fat in grams.
"carbohydrates" = amount of carbohydrates in grams.
"fibre" = amount of fibre in grams.
"all" = all nutrional information.}
}
\value{
A table of nutritional information for the food item.
}
\description{
\emph{get_nutrients}: get nutritional information for food items.
}
\examples{
get_nutrients("potato", "all") # get all nutritional information for all search hits of "potato"
get_nutrients("potato", "fibre")[1,] # get fibre information for the first search hit of "potato"

}
