% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfootprint_functions.R
\name{plot_wf_global}
\alias{plot_wf_global}
\title{\emph{plot_wf_global}: visualize the global water footprint of a food item.}
\usage{
plot_wf_global(food)
}
\arguments{
\item{food}{Any food item.}
}
\value{
A plot with the global blue, green, and grey water footprint of the food item.
}
\description{
\emph{plot_wf_global}: visualize the global water footprint of a food item.
}
\examples{
plot_wf_global(potato)

}
