% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nutrients.R
\name{search_food}
\alias{search_food}
\title{The function \emph{search_food} Searches for a specified foot item and returns a list of hits matching the search query, as well as a picture of the item.}
\usage{
search_food(food)
}
\arguments{
\item{food}{The food item, inputed as a string, that you would like to search for.}
}
\value{
A table of nutritional information for each search hit and a picture of the first hit.
}
\description{
The function \emph{search_food} Searches for a specified foot item and returns a list of hits matching the search query, as well as a picture of the item.
}
