% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfootprint_functions.R
\name{wf_country}
\alias{wf_country}
\title{\emph{wf_country}: retrieve the water footprint of a food item in a specific country.}
\usage{
wf_country(food, country)
}
\arguments{
\item{food}{Any food item.}

\item{country}{A two-letter country code according to the FIPS country codes.}
}
\value{
A table with the blue, green, and grey water footprint of the food item in that country.
}
\description{
\emph{wf_country}: retrieve the water footprint of a food item in a specific country.
}
\examples{
wf_country(potato, NL)

}
