% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfootprint_functions.R
\name{wf_country_plot}
\alias{wf_country_plot}
\title{\emph{wf_country_plot}: plot the water footprint of a food item across countries in a continent.}
\usage{
wf_country_plot(food, continent)
}
\arguments{
\item{food}{Any food item.}

\item{continent}{Specify the continent. Options: Americas, Africa, Asia, Oceania, Europe.}
}
\value{
A plot visualizing the blue, green, and grey water footprint of the food item per country in the specified continent.
}
\description{
\emph{wf_country_plot}: plot the water footprint of a food item across countries in a continent.
}
\examples{
wf_country_plot(potato, Europe)

}
