% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfootprint_functions.R
\name{wf_global}
\alias{wf_global}
\title{\emph{wf_global}: retrieve the global water footprint of a food item.}
\usage{
wf_global(food)
}
\arguments{
\item{food}{Any food item.}
}
\value{
A list, including a table with the global blue, green, and grey water footprint of the food item and nutritional information about the food item.
}
\description{
\emph{wf_global}: retrieve the global water footprint of a food item.
}
\examples{
wf_global(potato)

}
