% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfootprint_functions.R
\name{wf_region}
\alias{wf_region}
\title{\emph{wf_region}: retrieve the water footprint of a food item across regions of a country.}
\usage{
wf_region(food, country)
}
\arguments{
\item{food}{Any food item.}

\item{country}{A two-letter country code.}
}
\value{
A table of the blue, green, and grey water footprint of the food item per region in the specified country.
}
\description{
\emph{wf_region}: retrieve the water footprint of a food item across regions of a country.
}
\examples{
wf_region(potato, Netherlands)

}
