% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nutrients.R
\name{get_nutrients}
\alias{get_nutrients}
\title{The function \emph{get_nutrients} Returns a table of nutritional information for a specified food item.}
\usage{
get_nutrients(food, measure)
}
\arguments{
\item{food}{The food item, inputed as a string, you would like to search for.}

\item{measure}{The type of nutritional information, inputed as a string, per serving of the specified food item.
"calories" = The amount of calories (kcal).
"percent_nutrients" = the percentage of the daily needed intake of nutrients in the food item.
"fat" = amount of fat in grams.
"carbohydrates" = amount of carbohydrates in grams.
"fibre" = amount of fibre in grams.
"all" = all nutrional information.}
}
\value{
A table of nutritional information for the food item.
}
\description{
The function \emph{get_nutrients} Returns a table of nutritional information for a specified food item.
}
